added basic structure to user/chef dashboard pages, placed page files in folders
