+++ .env
@@ -1,0 +1,1 @@
+REACT_APP_GOOGLE_KEY=AIzaSyDjkiXSorawjm2A34CuENp-BWDCoubGy8w
