+++ package.json
@@ -4,8 +4,10 @@
   "private": true,
   "dependencies": {
     "dom": "0.0.3",
+    "dotenv": "^8.2.0",
     "react": "^16.11.0",
     "react-dom": "^16.11.0",
+    "react-google-maps": "^9.4.5",
     "react-router-dom": "^5.1.2",
     "react-scripts": "3.2.0",
     "router": "^1.3.3"
