+++ src/components/Container/Container.css
@@ -1,0 +1,3 @@
+.container {
+    margin-top: 60px;
+}
