+++ src/components/NavBar/NavBar.css
@@ -5,6 +5,7 @@
     letter-spacing: .1rem;
     background: rgba(0, 0, 0, 0.6)!important;
     /* background: transparent!important; */
+    transition: top 0.3s;
   }
   
   .navbar-brand:hover {
