+++ src/components/NavBar/NavBar.js
@@ -1,18 +1,27 @@
 import React, { Component } from 'react';
 import "./NavBar.css"
 
+var prevScrollpos = window.pageYOffset;
+window.onscroll = function () {
+    var currentScrollPos = window.pageYOffset;
+    if (prevScrollpos > currentScrollPos) {
+        document.getElementById("navbar").style.top = "0";
+    } else {
+        document.getElementById("navbar").style.top = "-60px";
+    }
+    prevScrollpos = currentScrollPos;
+}
 
 export default class NavBar extends Component {
     render() {
         return (
-            <nav className="navbar navbar-expand-md navbar-dark bg-dark fixed-top">
+            <nav id="navbar" className="navbar navbar-expand-md navbar-dark bg-dark fixed-top">
                 <a className="navbar-brand" href="/"><strong>OlafGo</strong></a>
-                <button className="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarResponsive"
-                    aria-controls="navbarResponsive" aria-expanded="false" aria-label="Toggle navigation">
+                <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarResponsive">
                     <span className="navbar-toggler-icon"></span>
                 </button>
                 <div className="collapse navbar-collapse" id="navbarResponsive">
-                    <ul className="navbar-nav mr-auto">
+                    <ul className="navbar-nav ml-auto">
                         <li className="nav-item active">
                             <a className="nav-link" href="/">Home<span className="sr-only">(current)</span></a>
                         </li>
