+++ src/components/pages/Chef/Chef.css
@@ -1,0 +1,3 @@
+.chef-dash {
+    padding-top: 50px;
+}
