+++ src/components/pages/Home/Home.css
@@ -73,7 +73,7 @@
   }
 
   .home-inner {
-    background-image: url("placeholder-head.jpg")
+    background-image: url("placeholder-head.jpg");
   }
 
   .caption {
