--- src/components/pages/User.js
@@ -1,14 +1,0 @@
-import React, { Component } from 'react';
-import NavBar from "../NavBar/NavBar";
-
-
-export default class User extends Component {
-    render() {
-        return (
-            <div>
-                <NavBar />
-               <h1>Here is the User page</h1> 
-            </div>
-        )
-    }
-}
