+++ src/components/pages/User/User.css
@@ -1,0 +1,3 @@
+.user-dash {
+    padding-top: 50px;
+}
